refactor: extract error-redirect logic from lib to server actions
