+++ app/_actions/index.ts
@@ -1,4 +1,5 @@
 export * from "./auth";
+export * from "./error-redirect";
 export * from "./flight-search-history";
 export * from "./flights";
 export * from "./orders";
