+++ app/_actions/session.ts
@@ -12,6 +12,18 @@ export interface SessionUser {
   image?: string | null;
 }
 
+export interface AuthResult {
+  success: boolean;
+  userId?: string;
+  error?: string;
+}
+
+function getLoginUrl(redirectTo?: string): string {
+  return redirectTo
+    ? `/auth/sign-in?redirect=${encodeURIComponent(redirectTo)}`
+    : "/auth/sign-in";
+}
+
 export async function getSessionUser(): Promise<SessionUser | null> {
   const headersList = await headers();
   const session = await auth.api.getSession({ headers: headersList });
@@ -34,11 +46,42 @@ export async function requireSessionUser(
   const user = await getSessionUser();
 
   if (!user) {
-    const loginUrl = redirectTo
-      ? `/auth/sign-in?redirect=${encodeURIComponent(redirectTo)}`
-      : "/auth/sign-in";
-    redirect(loginUrl);
+    redirect(getLoginUrl(redirectTo));
   }
 
   return user;
+}
+
+export async function getAuthenticatedUserId(): Promise<AuthResult> {
+  try {
+    const user = await getSessionUser();
+
+    if (!user?.id) {
+      return {
+        success: false,
+        error: "Authentication required. Please log in first.",
+      };
+    }
+
+    return {
+      success: true,
+      userId: user.id,
+    };
+  } catch (error) {
+    console.error("Authentication error:", error);
+    return {
+      success: false,
+      error: "Failed to verify authentication. Please try again.",
+    };
+  }
+}
+
+export async function requireAuth(redirectTo?: string): Promise<string> {
+  const authResult = await getAuthenticatedUserId();
+
+  if (!authResult.success || !authResult.userId) {
+    redirect(getLoginUrl(redirectTo));
+  }
+
+  return authResult.userId;
 }
